+++ NEWS.md
@@ -8,6 +8,7 @@ Red Panda C++ Version 1.0.4
   - fix: files will be saved to default encoding inspite of its original encoding
   - fix: parenthesis skip doesn't work when editing non-c/c++ files
   - enhancement: prefer local headers over system headers when complete #include header path
+  - fix: tab/shift+tab not correctly handled in options dialog's code template page
 
 Red Panda C++ Version 1.0.3
   - fix: when oj problem grabbed by competitive companion received,